refactor: Update group member model to use int64 for IDs
Updated the GroupMember model to use int64 for the ID, GroupID, and UserID fields. This change aligns the data types with the database schema and ensures consistency across the application, preventing potential data truncation or overflow issues.

+++ backend/internal/models/group_member.go
@@ -3,9 +3,9 @@ package models
 import "time"
 
 type GroupMember struct {
-	ID        int       `json:"id"`
-	GroupID   int       `json:"group_id"`
-	UserID    int       `json:"user_id"`
+	ID        int64     `json:"id"`
+	GroupID   int64     `json:"group_id"`
+	UserID    int64     `json:"user_id"`
 	Role      string    `json:"role"`
 	CreatedAt time.Time `json:"created_at"`
 }
